sys: get rid of pthreads.
Since C++11 threads are available on every platform that has threads, we
no longer need our pthread implementation. The LOL_FEATURE_CXX11_THREADS
macro is gone, too, and we now just use LOL_FEATURE_THREADS.

Note that it is still necessary to link with -lpthread or -pthread on
some platforms, so we don’t throw that part away.