build: drop support for SDL v1.
The only reason we had to keep this was for emscripten builds. We can
now use SDL v2 using “emconfigure ./configure CFLAGS="-s USE_SDL=2” and
a few additional flags.

SDL_Mixer v2 is not available in emscripten packages but either it will
be added at some time, or I will get rid of it altogether. An alternate
solution could be this code snippet I found:

    cd SDL2_mixer
    emconfigure ./configure --prefix=$(pwd)/dist --enable-music-ogg --disable-music-ogg-shared CFLAGS="-s USE_VORBIS=1"
    emmake make install