audio: refactor audio streaming mechanism.
Rename “channels” to “tracks” when talking about the mixing channels, to
avoid confusion with channels in the context of mono/stereo/5.1. Also allow
conversions between various formats and frequencies.

To do: allow to query the current audio device’s preferred frequency when
the client application is able to send the best possible data.